+++ CHANGELOG.md
@@ -7,6 +7,17 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [6.0.0] - 2024-03-05
+
+- Implements `deleteAllUserRoleAssociationsForRole`
+- Drops `(app_id, role)` foreign key constraint on `user_roles` table
+
+### Migration
+
+```sql
+ALTER TABLE user_roles DROP CONSTRAINT IF EXISTS user_roles_role_fkey;
+```
+
 ## [5.0.7] - 2024-02-19
 
 - Fixes vulnerabilities in dependencies